agent/requirements.txt
fastapi>=0.112.0
uvicorn>=0.30.0
openai>=1.52.0
python-dotenv>=1.0.1
pandas>=2.2.2
openpyxl>=3.1.5
jinja2>=3.1.4
matplotlib>=3.9.0
numpy>=2.0.0
requests>=2.32.3
python-multipart>=0.0.9
pydantic>=2.8.2
pytest>=8.3.0
